Group redux related files by feature instead of function
