+++ client/src/components/Admin/IssueForm/index.js
@@ -5,8 +5,8 @@ import DocumentTitle from 'react-document-title';
 import { RESOLUTION_TYPES } from 'common/actionTypes/voting';
 import Button from '../../Button';
 import Dialog from '../../Dialog';
-import { adminDeleteIssue, createIssue } from '../../../actionCreators/adminButtons';
-import { activeIssueExists, getIssue } from '../../../selectors/issues';
+import { adminDeleteIssue, createIssue } from '../../../features/adminButtons/actionCreators';
+import { activeIssueExists, getIssue } from '../../../features/issue/selectors';
 import Alternative from './Alternative';
 import Checkboxes from './Checkboxes';
 import SelectResolutionType from './SelectResolutionType';
