+++ client/src/components/App/VotingMenu/index.js
@@ -2,9 +2,9 @@ import React from 'react';
 import { connect } from 'react-redux';
 import Cookies from 'js-cookie';
 import { VOTING_NOT_STARTED, VOTING_IN_PROGRESS } from 'common/actionTypes/issues';
-import { submitAnonymousVote, submitRegularVote } from '../../../actionCreators/voting';
-import { getShuffledAlternatives } from '../../../selectors/alternatives';
-import { activeIssueExists, getIssue, getIssueId, getOwnVote, getIssueKey } from '../../../selectors/issues';
+import { submitAnonymousVote, submitRegularVote } from '../../../features/voting/actionCreators';
+import { getShuffledAlternatives } from '../../../features/alternative/selectors';
+import { activeIssueExists, getIssue, getIssueId, getOwnVote, getIssueKey } from '../../../features/issue/selectors';
 import Alternatives from '../../Alternatives';
 import Button from '../../Button';
 import css from './VotingMenu.css';
