+++ client/src/components/ConcludedIssueList/index.js
@@ -2,8 +2,8 @@ import React, { PropTypes } from 'react';
 import { connect } from 'react-redux';
 import ConcludedIssue from '../ConcludedIssue';
 import css from './ConcludedIssueList.css';
-import { getConcludedIssuesExceptLatest } from '../../selectors/issues';
-import { concludedIssueListIsEnabled } from '../../selectors/userSettings';
+import { getConcludedIssuesExceptLatest } from '../../features/issue/selectors';
+import { concludedIssueListIsEnabled } from '../../features/userSettings/selectors';
 
 function ConcludedIssueList({ concludedIssueListEnabled, issues }) {
   return (
