+++ client/src/components/LatestConcludedIssue/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import ConcludedIssue from '../ConcludedIssue';
-import { getLatestConcludedIssue } from '../../selectors/issues';
+import { getLatestConcludedIssue } from '../../features/issue/selectors';
 
 const LatestConcludedIssue = ({ issue }) => {
   if (!issue) {
