+++ client/src/features/sagas.js
@@ -1,7 +1,7 @@
 import createSagaMiddleware from 'redux-saga';
 import { all } from 'redux-saga/effects';
 
-import userSettings from './components/UserSettings/sagas';
+import userSettings from './userSettings/sagas';
 
 function* rootSaga() {
   yield all([userSettings()]);
