+++ client/src/features/userSettings/sagas.js
@@ -2,8 +2,8 @@ import { call, takeLatest, select } from 'redux-saga/effects';
 import { TOGGLE_NOTIFICATION } from 'common/actionTypes/userSettings';
 import { ENABLE_VOTING } from 'common/actionTypes/voting';
 import { notify, notifyPermission } from '../../utils/notification';
-import { getIssueText } from '../../selectors/issues';
-import { notificationIsEnabled } from '../../selectors/userSettings';
+import { getIssueText } from '../issue/selectors';
+import { notificationIsEnabled } from '../userSettings/selectors';
 
 function* openIssue() {
   const notificationEnabled = yield select(notificationIsEnabled);
